refactor : 이미지 presigned url 발급 메서드 분리
